requirements.txt
Flask==0.11.1
gunicorn==19.6.0
numpy==1.19.2 
easyocr-contrib-python-headless
requests
line-bot-sdk
